requirements.txt
adbutils
git+https://github.com/leng-yue/py-scrcpy-client@v0.5.0
joblib
Pillow
pydantic
pywin32
setuptools
toml